Display Tooltip when VM Template Descheduler not installed
Add the tooltip to the Descheduler to inform the user that "To enable the descheduler,
you must install the Kube Descheduler Operator from OperatorHub and enable one
or more descheduler profiles.", when the Descheduler is not installed, according
to the new design.

Remove the unnecessary check if the Descheduler is installed, for the ability of Button component
to be disabled when Descheduler operator not installed.

Update the Descheduler component's props:
use DeschedulerProps instead of TemplateSchedulingGridProps.

Rename useDeschedulerOn to isDeschedulerOn function, check only the appropriate annotation.