Bug 2101167: Fix the clickable area for Edit buttons in VM pages
Fix the area of Edit buttons in the VM Details and Scheduling tabs,
also in the Details, Scheduling and Metadata tabs/pages of VM creation flow.
Limit the clickable area only to the text and not the whole line,
so clicking on the empty space is not possible anymore.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2101167