Bug 2134668: Display correct command with username for SSH
This is the additional fix for Bug 2134668:
Get the correct cloud-init credentials even when the VM is stopped
and generate the ssh command properly (SSH over NodePort).
Disable the SSH over NodePort checkbox and display the tooltip to inform
the user in case that there is no username set.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2134668