Bug 2151056: Improve descriptive text of cloud-init and ssh-key
Improve the descriptive text of the Cloud-init field, to be more
precise, in th VM's Scripts tab, for both - already existing VM and the
VM that is going to be created.

For Authorized SSH Key field, remove "The key will be stored after the
machine is created" - for the VM that is already created.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2151056