Display only available DataSources for VM creation
Display only available DataSources when creating a new VM from a
bootable volume, instead of displaying all the DataSources,
and so fix the bug leading to not being able to create a VM or no being
able to start a VM - a VM created by selection of a volume that is not
ready/available (a.ka. DataSource with status condition type 'Ready'
and status 'False').

Sort the Instancetypes and preferences names in the appropriate drop
downs alphabetically, for better search experience.

Change 'Upload Data' to 'Upload data', also 'Preference' to 'preference'
in the 'Add bootable volume' modal, according to to design suggestions
and UI consistency.