CNV-18014: Update text in Add volume modal
Update the strings in the Add volume to boot from modal,
same ones in the Edit modal in the Bootable volumes list,
according to the design changes.
Also add missing helper texts to some of the modal's fields,
make the HelpIcon smaller to be displayed as expected.

Update also some of the strings in the same components
(as in Add volume to boot from modal) but used in other places
(e.g. PVC project and PVC name drop downs), for consistency reasons.

This PR is related to features:
https://issues.redhat.com/browse/CNV-18014
https://issues.redhat.com/browse/CNV-15501