Bug 2168486: Enable Restore template settings by fixing templates link
Make "Restore template settings" enabled, when editing VM's CPU/Memory,
by fixing incorrect link to the original Template the VM was created
from. Fix the incorrect namespace in the link, which was set to the
namespace of the VM when creating a VM, not the original Template's one.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2168486