Bug 2175983: Update text on Delete modal for bootable volumes
Update the title and the text of the Delete modal for bootable volumes:

Modal title text: "Delete volume parameters?"

Modal content: "Deleting the parameters will mark this volume
as non-bootable and remove it from the bootable volumes list.
The volume will still be available in the cluster."

This change of the text is made according to the design changes,
as the original text and the title of the modal weren't quite
appropriate or clear enough to the user.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2175983