Bug 2175986: Improve message for different StorageClass
Update the message displayed in "Add volume to boot from modal"
when uploading volume, for case that a different StorageClass is
selected than the default one, to the following one:

This StorageClass might cause slower cloning.

So the message is simplified and explains what a "poor performance"
(mentioned in the original message) means, when selecting
such a StorageClass.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2175986