Bug 2177578: Set columns' width for volumes in Catalog > InstanceTypes
Set the minimum width for each of the columns in the volumes list in
Catalog > InstanceTypes by using 'min-content' sizing keyword that
represents the intrinsic minimum width of the content.

Update TableData component to be able to get className as an optional
prop, to get styled according to requirements.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2177578