Remove extra buttons to create VM or MigrationPolicy
Don't display the unnecessary buttons to create VirtualMachine
or MigrationPolicy in the empty state pages (no resources present),
as there is the button to create the appropriate resource present
also in the center of the page, which is the default for Patternfly
empty state. Make this change according to the UX suggestions.