Bug 2183491: Truncate PVC name in Add volume modal
Truncate the PVC name options for the PVC name drop down in Add volume
modal (when selecting "Use existing volume" radio button) in case when
the name is too long according to the width of the modal. Display the
whole name in the popover displayed on mouseover.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2183491