Bug 2176797: Make "YAML" button consistent in Disks tabs
Move the Add buttons below the title of the page, add missing titles,
so the YAML switcher can be placed next to the title of the page,
to make the page more consistent with other pages containing YAML
swithcer. Make these changes in the following pages:
- VM Disks tab
- Catalog wizard Disks tab
- VM Template Disks tab

Refactor DiskListTitle component, update the related css files for
better look, make it universal, move it to utils and use it in the
pages mentioned above.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2176797