CNV-23128: Add OpenShift Virtualization Welcome Modal
Add the welcome modal to improve the UX for new users in the OpenShift
Virtualization space. The modal provides useful information such as the
most important/common actions users may take next and acts as shortcut.
It also helps users understand they just landed in the right place.

Currently the modal contains "Create VirtualMachine" button that leads
to Catalog page. It also contains the following links:
- "Create a virtual machine from a template (Quick start)" (leads to
  the appropriate Quick start),
- "Upload a boot source (Quick start)" (also leads to the appropriate
  Quick start).

Additionally, "Do not show this again" checkbox is displayed
in the modal, to provide ability not to show the modal to the user
again.

Fixes https://issues.redhat.com/browse/CNV-23128