CNV-29485: Move YAML switcher out of Template tabs
Move the YAML switcher out of the Templates tabs/pages,
display it to the left of Actions menu and only for relevant pages,
that contained the switcher also before this change.
Make this change according to the recent design updates.

Make some css updates for some pages to achieve more consistent look:
align sidebar editor properly in Disks and NICs tabs to be the same
as in the other pages (padding-top).

Fixes https://issues.redhat.com/browse/CNV-29485