CNV-29502: Move YAML switcher out of tabs in Review and create VM
Move the YAML switcher out of the tabs in Review and create
VirtualMachine page to the new place in the right - above the tabs,
same as it is for VM or Template tabs, for consistency reasons.
Display the switcher only for relevant pages that contained
the switcher also before this change. Make this change according
to the recent design updates.

Make some css updates for some pages to achieve more consistent look:
align sidebar editor properly in Disks and NICs tabs to be the same
as in the other pages (padding-top).

Fixes https://issues.redhat.com/browse/CNV-29502