CNV-29788: Remove titles from Environment, Disks tab for customizing VM
Remove titles from Environment and Disks tabs in "Review and create
VirtualMachine" to make the tabs more consistent with each other.
Add 'The following information is provided by the OpenShift
Virtualization operator.' text for Disks tab instead of help text icon
displying the popover containing this text, similarly as it is already
done in the Environment tab, to keep the information available to the
user.

Fixes https://issues.redhat.com/browse/CNV-29788