Bug 2221934: Update text for LoadBalancer section in Overview Settings
Update the name of the section from "LoadBalancer" to "LoadBalancer
service" to better reflect what the section is about. Update the
checkbox text to "Enable the creation of LoadBalancer services for SSH
connections to VirtualMachines", subtext to "A load balancer must be
configured".

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2221934