CNV-32531: Fix error in Catalog drawer for "All Projects"
"Error: the server could not find the requested resource" appeared
in Catalog drawer for "All Projects" Project. This commit fixes the
error by properly setting the namespace for the template object
that needs to be processed before moving forward in VM creation.

The root cause of the problem was that the namespace in the code
for "All Projects" was "undefined", so the correct template object
couldn't be created and then processing such an incorrect template
object failed and lead to the mentioned error.

Fixes https://issues.redhat.com/browse/CNV-32531