CNV-32609: Keep the CPU value from Catalog drawer for customizing page
Save correctly the CPU value from Catalog drawer, especially if the user
has edited this value in the drawer, so it can be displayed in Customize
and create VirtualMachine page correctly, same as it was set when the
user has left the drawer.

Additionally, add missing types in some related code (such as
V1VirtualMachine or TabsData types).

Fixes https://issues.redhat.com/browse/CNV-32609