CNV-32787: Adjust length of VM name field in Catalog drawer
Sometimes, the generated VM name is a bit longer, so the VM name field
cannot hold it completely. This commit makes it a bit longer to hold the
long name, as there's enough space in the line of the drawer.

Additionally, edit the padding of the related part of the drawer the way
that the whole content of the Catalog drawer is nicely aligned.

Fixes https://issues.redhat.com/browse/CNV-32787