Make running yarn dev with no errors
Fix multiple errors occurred after running "yarn dev", and showing
mainly the following:
"<ComponentName> cannot be used as a JSX component"

Fix the errors by adding missing "@types/react-dom" package, adding
"@patternfly/react-icons" and "@types/react-dom" to the resolutions in
the package.json file, as some of our packages use those packages as
theirdependencies, too, and that lead to the mess as yarn does not do
deduplication and those packages were installed more times, which lead
to typescript being confused. And so the following line
"react": ["./node_modules/react", "./node_modules/@types/react"]
had to be added to tsconfig.json file, too, to tell it explicitly,
which types for react to use.