CNV-37383: Move "Not available" above "Dynamic SSH key injection .." msg
Move "Not available" text (or the name of the existing SSH key if
present) above the message "Dynamic SSH key injection is not enabled
in this virtual machine. Learn more" in the VM's Configuration SSH tab
to make the information more clear to the user. Make this change
according to the UX decisions.

Fixes https://issues.redhat.com/browse/CNV-37383