CNV-37818: Set white background to all PageSection component occurrences
Set the background for Patternfly PageSection component we use back to
white due to the change of the default setting of the component from
white to grey. Adjust the component by adding
"variant={PageSectionVariants.light}" to achieve the expected look
in the various pages.

Fixes https://issues.redhat.com/browse/CNV-37818