Add Sysprep option for Windows VMs when creating from InstanceTypes
Adjust the third section of the form when creating VMs from
InstanceTypes depending on the VM type that is being created:
Linux or Windows (according to the instancetype preference selection).
Add Sysprep option for Windows VMs instead of SSH data to the section.

Fixes https://issues.redhat.com/browse/CNV-35740