fix: Clear revisionName when updating instance type
While this behaviour is likely to change in the near future at present
users looking to move between instance types need to clear the
revisionName to ensure the correct LiveUpdate logic kicks in. While this
is called out by the user-guide docs the current wording doesn't make
this obvious and will be updated [1].

This commit ensures the UI clears the revisionName field when updating
an instance type ensuring that LiveUpdate and any associated hot plug
logic is activated correctly.

[1] https://kubevirt.io/user-guide/user_workloads/instancetypes/#versioning

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>