Initialize interface type based only on the network interface
Originally the interface type was initialized based on list of
NetworkAttachmentDefinition objects retrieved dynamically (see
CNV-31299). With changes introduced in CNV-43742 and CNV-47511 the
approach switched to using the interface. At this point it's no longer
needed to place the initialization logic in the useEffect() - it can be
moved to the parent component.

Note that we have now a proxy wrapper around interfacesTypes mapping.
The fallback code to set interface type in NetworkInterfaceModal is
not used. The proxy will default to bridge if there is no mapping.

Reference-Url: https://github.com/kubevirt-ui/kubevirt-plugin/pull/1424
Reference-Url: https://github.com/kubevirt-ui/kubevirt-plugin/pull/2155
Reference-Url: https://github.com/kubevirt-ui/kubevirt-plugin/pull/2076
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>