Use default boot mode from the preference
Fixed issues:
1. UEFI mode used as default boot mode in the customization details.
   Fixed by strict check in getPreferredBootmode()
2. BIOS mode used as default mode (regardless of preference) in the edit
   modal. Fixed by passing the preferred boot mode (already available).
3. pending changes warning displayed after creating a VM with
   preference that includes boot mode. Fixed by detecting that no
   override was provided in the VM resource.

Additionaly, the preference resolving policy was changed. Before, the
default was VirtualMachinePreference (if cluster preference was not
matched). After this PR, the default is VirtualMachineClusterPreference.
The reason for this change: some of the data source volumes are not
providing the instancetype.kubevirt.io/default-preference-kind label
which blocked resolving the preference (silently) on the test envs.
As a result incorrect data was presented in the customization details -
including incorrect boot mode. In this case the cluster preference seems
a better fallback.

Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>