Detect single node cluster mode based on number of nodes
The single node cluster node detection is used in the UI to block
live migrations between nodes. Before this PR the check was based on
Infrastructure CR and infrastructureTopology property. If the property
was `SingleReplica` then single cluster node was assumed.

After, this PR the check will be based on actual number of nodes
in the cluster.

Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>