Drop detecting single node cluster mode
Before, the single node cluster node detection was used in the UI to
block live migrations between nodes. The check was based on
Infrastructure CR and infrastructureTopology property. If the property
was `SingleReplica` then single cluster node was assumed.
This assumption was incorrect as non high-availability cluster may
consist of more than one node.

After dropping single cluster check it will be up to the cluster to
decide if live migration is allowed. The insufficient number of nodes
will be treated in the same way as missing resources.

Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>