Separate detected pending changes from conditions set on the VM
Before this PR the UI first detected the changes (i.e. added NIC) and
then tried to guess what action is needed to make the change persistent.
Possible choices were: manual restart or manual live migration.

With CNV-58395 the RestartRequired condition is set for all changes
that are classified as not hot-pluggable. In 4.20 also the LiveUpdate
vmRolloutStrategy is the default which in theory removes the need of
manual live migration. However after this improvements there are still
some open points:
1. RestartRequired or MigrationRequired conditions are not linked with
   a particular change and are added  with a delay.
2. Detection logic in the UI might be outdated or incorrect.
3. The cluster may use vmRolloutStrategy Stage
4. automatic migration might by stuck for a longer time

In this PR the above points are addressed:
1. RestartRequired condition is always shown (regardless of the detected
   changes)
2. detected pending changes in networking area are always shown but
   without any suggested action. It's purely information that VM and
   VMI have different spec.
3. if cluster is using LiveUpdate strategy the information about pending
   changes will automatically disappear after the changes are applied.
4. if the cluster is using Stage strategy or if the automatic migration
   is blocked then the user will need to handle the changes manually

Fixed scenarios:
1. changing NetworkAttachementDefinition - CNV-58657
2. changing network model i.e. virtio -> e1000e
3. all hot-plug changes (adding bridge/SR-IOV, removing bridge)

Reference-Url: https://issues.redhat.com/browse/CNV-58657
Reference-Url: https://issues.redhat.com/browse/CNV-58395
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>