Detect concurrent conflicting NIC updates
The network and interface section of the VM config are modified by the
backend i.e. during hot-plug flow.
Before, the UI could overwrite backgtound modifications by patching the
config without checking if the data changed in the meantime.

After this PR, a PATCH test operation will be used before network and
interface replace/remove operations to ensure the resource didn't change.
Additionally the updates will be limited to a single NIC - not the
entire array of networks/interfaces.

Reference-Url: https://datatracker.ietf.org/doc/html/rfc6902#section-4.6
Reference-Url: https://github.com/kubernetes/kubernetes/issues/70281
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>